requirements.txt
# ── Core ─────────────────────────────────────────────────────────────────────
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic[email]==2.10.3
python-multipart==0.0.18

# ── Database ──────────────────────────────────────────────────────────────────
sqlalchemy==2.0.36

# ── Auth ─────────────────────────────────────────────────────────────────────
bcrypt==4.2.1
python-jose[cryptography]==3.3.0

# ── AI / Data ─────────────────────────────────────────────────────────────────
pandas==2.2.3
scikit-learn==1.5.2
numpy==1.26.4

# ── Frontend ─────────────────────────────────────────────────────────────────
streamlit==1.40.2
requests==2.32.3

# ── Testing ───────────────────────────────────────────────────────────────────
pytest==8.3.4
httpx==0.28.0
